Current Build
I have added in a few menus (main menu and lose screen), The game is rather bland in terms of looks.

I have updated the build file and could do with some unit tests to ensure playability.

+++ GameCode/PROCJAM 2019/PROCJAM 2019/src/GameWindow.cpp
@@ -93,8 +93,8 @@ sf::Vector2f GameWindow::m_GetLowerBounds()
 {
 	sf::Vector2f l_ReturnVector;
 
-	l_ReturnVector = sf::Vector2f(m_RenderWindow.getPosition().x + m_RenderWindow.getSize().x, 
-		m_RenderWindow.getPosition().y + m_RenderWindow.getSize().y);
+	l_ReturnVector = sf::Vector2f((float)m_RenderWindow.getPosition().x + (float)m_RenderWindow.getSize().x, 
+		(float)m_RenderWindow.getPosition().y + (float)m_RenderWindow.getSize().y);
 
 	return l_ReturnVector;
 }
